Rename component to widget, now that we really have a mix of both.
