+++ Widgets/BestExitComparisonInfo.py
@@ -1,11 +1,11 @@
 import tkinter as tk
-from Components import Info
+from Widgets import Info
 from util import timeHelpers as timeh
 
-class GoldDiffInfo(Info.Info):
+class BestExitComparisonInfo(Info.Info):
     def __init__(self,parent,state,config):
         Info.Info.__init__(self,parent,state,config)
-        self.header.configure(text="Last Split (vs Best):")
+        self.resetUI()
 
     def hide(self):
         self.info.configure(text="-",fg=self.config["colours"]["text"])
@@ -16,16 +16,17 @@ class GoldDiffInfo(Info.Info):
         if self.state.splitnum and self.shouldHide():
             self.hide()
             return
-        if not timeh.greater(self.state.comparisons[0].segments[self.state.splitnum],self.state.segmentTime)\
-            and not timeh.isBlank(self.state.comparisons[0].segments[self.state.splitnum])\
-            and not (self.state.splitnum and timeh.isBlank(self.state.currentRun.segments[self.state.splitnum-1])):
+        if (not timeh.greater(self.state.comparisons[0].segments[self.state.splitnum],self.state.segmentTime)\
+            and not timeh.isBlank(self.state.comparisons[0].segments[self.state.splitnum]))\
+            or (not timeh.greater(self.state.comparisons[3].totals[self.state.splitnum],self.state.totalTime)\
+            and not timeh.isBlank(self.state.comparisons[3].totals[self.state.splitnum]))\
+            and not (self.state.splitnum and timeh.isBlank(self.state.currentRun.totals[self.state.splitnum-1])):
 
-            if self.state.splitnum and self.shouldHide():
+            if self.shouldHide():
                 self.hide()
                 return
 
-            self.header.configure(text="Current Segment:")
-            self.setTimes(self.state.segmentTime,previous=False)
+            self.setTimes(self.state.totalTime,previous=False)
 
     def onSplit(self):
         self.splitEndUpdate()
@@ -38,23 +39,23 @@ class GoldDiffInfo(Info.Info):
             if self.shouldHide():
                 self.hide()
                 return
-            self.setTimes(self.state.currentRun.segments[self.state.splitnum-1])
+            self.setTimes(self.state.currentRun.totals[self.state.splitnum-1])
 
     def onRestart(self):
         self.resetUI()
 
     def resetUI(self):
-        self.header.configure(text="Last Split (vs Best):")
+        self.header.configure(text="Vs Best Exit:")
         self.info.configure(text="")
 
     def splitEndUpdate(self):
         if not self.state.splitnum:
             return
-        self.header.configure(text="Last Split (vs Best):")
         if self.shouldHide():
             self.hide()
             return
-        self.setTimes(self.state.currentRun.segments[self.state.splitnum-1])
+        self.header.configure(text="Vs Best Exit:")
+        self.setTimes(self.state.currentRun.totals[self.state.splitnum-1])
 
     def setTimes(self,time,previous=True):
         if previous:
@@ -65,7 +66,7 @@ class GoldDiffInfo(Info.Info):
             timeh.timeToString(\
                 timeh.difference(\
                     time,\
-                    self.state.comparisons[0].segments[splitnum]\
+                    self.state.comparisons[3].totals[splitnum]\
                 ),\
                 {\
                     "showSign": True,\
@@ -84,23 +85,33 @@ class GoldDiffInfo(Info.Info):
         if timeh.isBlank(self.state.currentComparison.segments[split]):
             return self.config["colours"]["skipped"]
 
-        elif timeh.greater(self.state.currentComparison.segments[split],self.state.segmentTime):
-            return self.config["colours"]["gaining"]
         else:
-            return self.config["colours"]["losing"]
+            return self.setColour(self.state.segmentTime,self.state.totalTime,split)
 
     def setPreviousColour(self):
         split = self.state.splitnum-1
-        if timeh.isBlank(self.state.comparisons[0].segments[split])\
-            or timeh.isBlank(self.state.currentRun.segments[split])\
-            or timeh.isBlank(self.state.currentComparison.segments[split]):
+        if timeh.isBlank(self.state.comparisons[3].totals[split])\
+            or timeh.isBlank(self.state.currentRun.totals[split])\
+            or timeh.isBlank(self.state.currentComparison.totals[split]):
             return self.config["colours"]["skipped"]
 
-        if timeh.greater(self.state.comparisons[0].segments[split],self.state.currentRun.segments[split]):
+        if timeh.greater(self.state.comparisons[3].totals[split],self.state.currentRun.totals[split]):
             return self.config["colours"]["gold"]
+        else:
+            return self.setColour(self.state.currentRun.segments[split],self.state.currentRun.totals[split],split)
 
-        elif timeh.greater(self.state.currentComparison.segments[split],self.state.currentRun.segments[split]):
-            return self.config["colours"]["gaining"]
+    def setColour(self,segment,total,split):
+        if timeh.greater(self.state.comparisons[3].totals[split],total):
+
+            if timeh.greater(self.state.comparisons[3].segments[split],segment):
+                return self.config["colours"]["aheadGaining"]
+
+            else:
+                return self.config["colours"]["aheadLosing"]
 
         else:
-            return self.config["colours"]["losing"]
+            if timeh.greater(self.state.comparisons[3].segments[split],segment):
+                return self.config["colours"]["behindGaining"]
+
+            else:
+                return self.config["colours"]["behindLosing"]
