+++ Widgets/Info.py
@@ -1,12 +1,12 @@
 import tkinter as tk
-from Components import Component
+from Widgets import WidgetBase
 
-class Info(Component.Component):
+class Info(WidgetBase.WidgetBase):
     header = None
     info = None
 
     def __init__(self,parent,state,config):
-        Component.Component.__init__(self,parent,state,config)
+        WidgetBase.WidgetBase.__init__(self,parent,state,config)
         self.configure(bg=config["colours"]["bg"],padx=state.config["padx"])
         self.header = tk.Label(self, fg=config["colours"]["text"], bg=config["colours"]["bg"])
         self.info = tk.Label(self, fg=config["colours"]["text"], bg=config["colours"]["bg"])
