+++ Widgets/Timer.py
@@ -1,12 +1,12 @@
 import tkinter as tk
-from Components import Component
+from Widgets import WidgetBase
 from util import timeHelpers as timeh
 
-class Timer(Component.Component):
+class Timer(WidgetBase.WidgetBase):
     main = None
 
     def __init__(self,parent,state,config):
-        Component.Component.__init__(self,parent,state,config)
+        WidgetBase.WidgetBase.__init__(self,parent,state,config)
         self.configure(bg=config["colours"]["bg"],padx=state.config["padx"])
         self.main = tk.Label(self, bg=config["colours"]["bg"], fg=config["colours"]["main"], font=config["font"])
         m = config["position"]
